code refactor: resync Base when Manifest/HelmChart changed
